pyproject.toml
[tool.poetry]
name = "nome"
version = "0.1.0"
description = ""
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.0"
matplotlib = "^3.7.2"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 